feat(gateway): serve Your Throttle per-tenant on hosted; retire the #1848 deny
The Stats / Your Throttle dashboard was denied (404) on the hosted Gateway because the
feed was described as fleet-global. The aggregator behind it is now tenant-partitioned
(MTR-08 - every map keyed by (tenant, ...), every read takes a TenantId), so there is a
correct per-tenant answer to serve.

- Retire the hosted deny (the group endpoint filter is gone).
- GET /stats/data resolves the CALLER's tenant from its authenticated device key
  (ResolveReadTenant, the same seam the served settings routes use) and serves only that
  tenant's totals. A request with no resolvable tenant is refused 403 - NEVER served the
  Local partition. Self-host resolves to the single Local tenant, unchanged.
- The display time zone is read for the caller's tenant rather than the process global.
- The embedded /stats HTML shell serves on both surfaces; it carries no per-tenant data
  and the feed it fetches is what is tenant-gated.

Tests: HostedStatsServeTests (real hosted GatewayHost, two enrolled tenants + one unbound
device) proves serve for an enrolled tenant, 403 for an unresolved tenant, and that one
tenant's turns/repos are invisible to another. The self-host control proves the page and
feed still serve their real totals with hosted mode explicitly off, in both non-hosted forms.
